Thread IcebergMetricsConfig through SerializeValue; move metrics types to their own header
Addresses review on #1139:
- Move IcebergMetricsMode / IcebergMetricsConfig and the metrics-mode resolvers
  out of iceberg_insert.cpp into core/expression/iceberg_metrics.{hpp,cpp}.
- Thread IcebergMetricsConfig through the generic SerializeValue instead of a
  string-only SerializeStringBound helper (removed). The VARCHAR case truncates
  using metrics_config.truncate_length; other types are unaffected. Defaults to
  truncate(16), so existing callers (e.g. partition summaries) are unchanged.
